Creates separate component for SpecificAssetIds
